ammo and flotsam (#1454)
* change Flosam::UnitSize() to use double

This change allows UnitSize to accurately describe the size of what it contains since some items have fractional masses.

This fixes ships trying to pick up flotsam they don't have room for, typically ammo

* change Ship::Jettison outfit to allow non-full flotsam boxes

This is to stop fractional mass items from disappearing if they don't completely fill a flotsam box.

removed static keyword to allow box sizes to be appropriate for each item

+++ source/Flotsam.h
@@ -58,7 +58,7 @@ public:
 	int Count() const;
 	// This is how big one "unit" of the flotsam is (in tons). If a ship has
 	// less than this amount of space, it can't pick up anything here.
-	int UnitSize() const;
+	double UnitSize() const;
 	
 	
 private:
